Create a testing-server to serve the mock data in development environment (#868)
In this patch, we created a testing-server as the mock go-api server which serve mock data for the sake of testing requirement. Running npm script `npm run start-testing-server` returns mock data with several endpoints for multiple pages listed as below : 

* homepage
* listed-post pages (topic pages, photography pages ...etc)
* author-list page
* author-collection pages
* landing pages of topics which can be linked from homepage 
* articles including related articles which can be linked from homepage

Note that before starting the start-testing-server, it would automatically check whether 8080 port is used and show warning message then exit if the port is already taken.